Fix: only proceed with TF/Ansible reconfigure for terminated instances
If the user passes passes to the --nodes option a non-existing hostname
to be remove, then instance termination will fail and this is reported.
Nevertheless the script carries on trying to remove this non-existing
host from the slurm configuration, which might even in some cases succeed.

This change updates the list of hosts to remove based on whether the
instance termination was processed without an error.